sidecar/pyproject.toml
[project]
name = "ouro-ops-runner"
version = "0.1.0"
description = "Ansible runner bridge for Ouro Ops"
requires-python = ">=3.11"
dependencies = [
    "ansible-runner>=2.4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = []

[tool.uv]
dev-dependencies = []
